requirements.txt
python-binance==1.0.17
python-dotenv==1.0.0